Merge bug-fix-#216 to develop

+++ src/ASP.NET/GoAber/GoAber/Areas/Admin/Controllers/ApplicationUsersController.cs
@@ -34,6 +34,19 @@ namespace GoAber.Areas.Admin.Controllers
             var teamList = teamsService.CreateTeamList();
             ViewBag.TeamId = new SelectList(teamList, "TeamId", "Name", "CommunityName", 0);
 
+            // CG - Use the ASP.NET 'TempData' dictionary to pass a value IMMEDIATELY from a re-direct back to this controller. (Set in the 'Delete' action)
+            // See: http://stackoverflow.com/a/15958277 for more information.
+            if (TempData != null)
+            {
+                if (TempData.ContainsKey("Error") && TempData["Error"] != null)
+                {
+                    ViewBag.FailureMessage = TempData["Error"].ToString();
+                    TempData["Error"] = null;
+                }
+                
+            }
+           
+
             return View(applicationUsers.ToPagedList(pageNumber, pageSize));
         }
 
@@ -47,6 +60,8 @@ namespace GoAber.Areas.Admin.Controllers
 
             var teamList = teamsService.CreateTeamList();
             ViewBag.TeamId = new SelectList(teamList, "TeamId", "Name", "CommunityName", 0);
+
+            ViewBag.ConnorMessage = "The Same test Type might have been already created,, go back to the Visit page to see the avilalbe Lab Tests";
 
             return View(applicationUsers.ToPagedList(1, pageSize));
         }
@@ -191,9 +206,37 @@ namespace GoAber.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
-            ApplicationUser applicationUser = db.Users.Find(id);
-            db.Users.Remove(applicationUser);
-            db.SaveChanges();
+
+            // CG - Prevent an administrator user from deleting their own account whilst still logged in.
+            if (ApplicationUserService.GetCurrentApplicationUser().Id.Equals(id))
+            {
+                TempData["Error"] = "You are currently attempting to delete your own account whilst still logged in. If you wish to delete this account, please sign out and log in using another administrator account.";
+
+            }
+            else
+            {
+                ApplicationUser applicationUser = db.Users.Find(id);
+
+                // CG - If we want to delete an ApplicationUser, we need to make sure that we remove all of the FK relationships.
+                if (applicationUser != null)
+                {
+                    db.ActivityDatas.RemoveRange(db.ActivityDatas.Where(u => u.ApplicationUserId.Equals(applicationUser.Id)));
+
+                    db.Audit.RemoveRange(db.Audit.Where(u => u.ApplicationUserId.Equals(applicationUser.Id)));
+
+                    db.DataRemovalAudits.RemoveRange(db.DataRemovalAudits.Where(u => u.ApplicationUserId.Equals(applicationUser.Id)));
+
+                    db.WebServiceAuths.RemoveRange(db.WebServiceAuths.Where(u => u.ApplicationUserId.Equals(applicationUser.Id)));
+
+                    db.Users.Remove(applicationUser);
+
+                    db.SaveChanges();
+                }
+                
+
+            }
+
+            
             return RedirectToAction("Index");
         }
